requirements.txt
streamlit==1.28.1
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.4.0
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0